+++ examples/blueprint_example.py
@@ -18,4 +18,4 @@ def index(request):
 
 
 if __name__ == "__main__":
-    app.run("127.0.0.1", port=8098, debug=True)
+    app.run("127.0.0.1", port=8098, debug=True, auto_reload=False)