+++ spf/context.py
@@ -7,14 +7,29 @@ own children.
 """
 
 
-class ContextDict(object):
+class HierDict(object):
     """
 This is the specialised dictionary that is used by Sanic Plugins Framework
 to manage Context objects. It can be hierarchical, and it searches its
 parents if it cannot find an item in its own dictionary. It can create its
 own children.
     """
-    __slots__ = ('_spf', '_parent_context', '_dict', '__weakref__')
+    __slots__ = ('_parent_hd', '_dict', '__weakref__')
+
+    @classmethod
+    def _iter_slots(cls):
+        use_cls = cls
+        bases = cls.__bases__
+        base_count = 0
+        while True:
+            if use_cls.__slots__:
+                for _s in use_cls.__slots__:
+                    yield _s
+            if base_count >= len(bases):
+                break
+            use_cls = bases[base_count]
+            base_count += 1
+        return
 
     def _inner(self):
         """
@@ -25,11 +40,11 @@ own children.
 
     def __repr__(self):
         _dict_repr = repr(self._inner())
-        return "ContextDict({:s})".format(_dict_repr)
+        return "HierDict({:s})".format(_dict_repr)
 
     def __str__(self):
         _dict_str = str(self._inner())
-        return "ContextDict({:s})".format(_dict_str)
+        return "HierDict({:s})".format(_dict_str)
 
     def __len__(self):
         return len(self._inner())
@@ -43,16 +58,16 @@ own children.
             return self._inner().__getitem__(item)
         except KeyError as e1:
             parents_searched = [self]
-            parent = self._parent_context
+            parent = self._parent_hd
             while parent:
                 try:
                     return parent._inner().__getitem__(item)
                 except KeyError:
                     parents_searched.append(parent)
                     # noinspection PyProtectedMember
-                    next_parent = parent._parent_context
+                    next_parent = parent._parent_hd
                     if next_parent in parents_searched:
-                        raise RuntimeError("Recursive ContextDict found!")
+                        raise RuntimeError("Recursive HierDict found!")
                     parent = next_parent
             raise e1
 
@@ -60,7 +75,7 @@ own children.
         self._inner().__delitem__(key)
 
     def __getattr__(self, item):
-        if item in self.__slots__:
+        if item in self._iter_slots():
             return object.__getattribute__(self, item)
         try:
             return self.__getitem__(item)
@@ -68,7 +83,7 @@ own children.
             raise AttributeError(*e.args)
 
     def __setattr__(self, key, value):
-        if key in self.__slots__:
+        if key in self._iter_slots():
             if key == '__weakref__':
                 if value is None:
                     return
@@ -98,7 +113,7 @@ own children.
 
     def items(self):
         """
-        A set-like read-only view ContextDict's (K,V) tuples
+        A set-like read-only view HierDict's (K,V) tuples
         :return:
         :rtype: frozenset
         """
@@ -106,7 +121,7 @@ own children.
 
     def keys(self):
         """
-        An object containing a view on the ContextDict's keys
+        An object containing a view on the HierDict's keys
         :return:
         :rtype: tuple  # using tuple to represent an immutable list
         """
@@ -114,7 +129,7 @@ own children.
 
     def values(self):
         """
-        An object containing a view on the ContextDict's values
+        An object containing a view on the HierDict's values
         :return:
         :rtype: tuple  # using tuple to represent an immutable list
         """
@@ -122,9 +137,9 @@ own children.
 
     def replace(self, key, value):
         """
-        If this ContextDict doesn't already have this key, it sets
-        the value on a parent ContextDict if that parent has the key,
-        otherwise sets the value on this ContextDict.
+        If this HierDict doesn't already have this key, it sets
+        the value on a parent HierDict if that parent has the key,
+        otherwise sets the value on this HierDict.
         :param key:
         :param value:
         :return: Nothing
@@ -133,7 +148,7 @@ own children.
         if key in self._inner().keys():
             return self.__setitem__(key, value)
         parents_searched = [self]
-        parent = self._parent_context
+        parent = self._parent_hd
         while parent:
             try:
                 if key in parent.keys():
@@ -144,14 +159,14 @@ own children.
             # noinspection PyProtectedMember
             next_parent = parent._parent_context
             if next_parent in parents_searched:
-                raise RuntimeError("Recursive ContextDict found!")
+                raise RuntimeError("Recursive HierDict found!")
             parent = next_parent
         return self.__setitem__(key, value)
 
     # noinspection PyPep8Naming
     def update(self, E=None, **F):
         """
-        Update ContextDict from dict/iterable E and F
+        Update HierDict from dict/iterable E and F
         :return: Nothing
         :rtype: None
         """
@@ -168,30 +183,28 @@ own children.
         for K in F:
             self.replace(K, F[K])
 
-    def create_child_context(self, *args, **kwargs):
-        return ContextDict(self._spf, self, *args, **kwargs)
-
-    def __new__(cls, spf, parent, *args, **kwargs):
-        self = super(ContextDict, cls).__new__(cls)
+    def __new__(cls, parent, *args, **kwargs):
+        self = super(HierDict, cls).__new__(cls)
         self._dict = dict(*args, **kwargs)
         if parent is not None:
-            assert isinstance(parent, ContextDict),\
-                "Parent context must be a valid initialised ContextDict"
-            self._parent_context = parent
+            assert isinstance(parent, HierDict),\
+                "Parent context must be a valid initialised HierDict"
+            self._parent_hd = parent
         else:
-            self._parent_context = None
-        self._spf = spf
+            self._parent_hd = None
         return self
 
     def __init__(self, *args, **kwargs):
         args = list(args)
         args.pop(0)  # remove spf
         args.pop(0)  # remove parent
-        super(ContextDict, self).__init__()
+        super(HierDict, self).__init__()
 
     def __getstate__(self):
         state_dict = {}
-        for s in self.__slots__:
+        for s in HierDict.__slots__:
+            if s == "__weakref__":
+                continue
             state_dict[s] = object.__getattribute__(self, s)
         return state_dict
 
@@ -202,6 +215,42 @@ own children.
     def __reduce__(self):
         state_dict = self.__getstate__()
         spf = state_dict.pop('_spf')
-        parent_context = state_dict.pop('_parent_context')
-        return (ContextDict.__new__, (self.__class__, spf, parent_context),
+        parent_context = state_dict.pop('_parent_hd')
+        return (HierDict.__new__, (self.__class__, spf, parent_context),
+                state_dict)
+
+
+class SanicContext(HierDict):
+    __slots__ = ('_spf',)
+
+    def __repr__(self):
+        _dict_repr = repr(self._inner())
+        return "SanicContext({:s})".format(_dict_repr)
+
+    def __str__(self):
+        _dict_str = str(self._inner())
+        return "SanicContext({:s})".format(_dict_str)
+
+    def create_child_context(self, *args, **kwargs):
+        return SanicContext(self._spf, self, *args, **kwargs)
+
+    def __new__(cls, spf, parent, *args, **kwargs):
+        if parent is not None:
+            assert isinstance(parent, SanicContext),\
+                "Parent context must be a valid initialised SanicContext"
+        self = super(SanicContext, cls).__new__(cls, parent, *args, **kwargs)
+        self._spf = spf
+        return self
+
+    def __getstate__(self):
+        state_dict = super(SanicContext, self).__getstate__()
+        for s in SanicContext.__slots__:
+            state_dict[s] = object.__getattribute__(self, s)
+        return state_dict
+
+    def __reduce__(self):
+        state_dict = self.__getstate__()
+        spf = state_dict.pop('_spf')
+        parent_context = state_dict.pop('_parent_hd')
+        return (SanicContext.__new__, (self.__class__, spf, parent_context),
                 state_dict)