+++ spf/plugins/__init__.py
@@ -1,1 +1,4 @@
 # -*- coding: utf-8 -*-
+#
+from .contextualize import Contextualize, contextualize
+